requirements.txt
celery>=5.2.7,<5.3
django>=4.1.7,<4.2
djangorestframework>=3.14.0,<3.15
drf-spectacular>=0.25.1,<0.26
psycopg2>=2.9.5,<2.10
redis>=4.5.1,<4.6
